LayoutTests/imported/w3c/web-platform-tests/navigation-api/ordering-and-transition/reload-intercept-reject_no-currententrychange-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: boo

Harness Error (FAIL), message = Unhandled rejection: boo

PASS event and promise ordering for navigation.reload() intercepted by intercept()
